pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kabu-per-bot"
version = "0.1.0"
description = "Stock watch bot for PER/PSR notifications"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
]

[project.optional-dependencies]
gcp = [
  "google-cloud-firestore>=2.16.0",
  "firebase-admin>=6.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]
